services/agent_service.py
import uuid
import time
from typing import Dict, Optional
from datetime import datetime
from models.conversion_job import ConversionJob, JobStatus
from models.nifi_flow import Notebook
from services.nifi_parser import NiFiParser


class AgentService:
    """
    Service for invoking the agentic conversion system.

    NOTE: This is a placeholder implementation. The actual agent integration
    will depend on the team's agent API design (REST API, notebook, MLflow, etc.)
    """

    def __init__(self):
        # In-memory job storage (replace with database in production)
        self.jobs: Dict[str, ConversionJob] = {}

    def start_conversion(self, nifi_xml: str, source_file: str, source_volume_path: str, user_id: str) -> ConversionJob:
        """
        Start the conversion process for a NiFi XML file.

        Args:
            nifi_xml: The NiFi XML content
            source_file: The source file name
            source_volume_path: The volume path of the source file
            user_id: The user initiating the conversion

        Returns:
            ConversionJob object with job details
        """
        job_id = str(uuid.uuid4())

        job = ConversionJob(
            job_id=job_id,
            user_id=user_id,
            source_file=source_file,
            source_volume_path=source_volume_path,
            status=JobStatus.PENDING,
            progress_percentage=0,
            status_message="Job queued for processing"
        )

        self.jobs[job_id] = job

        # TODO: Replace with actual agent invocation
        # This is where you would:
        # 1. Call the agent API/service
        # 2. Pass the NiFi XML content
        # 3. Get back a job/task ID for tracking

        # For now, simulate the process
        self._simulate_conversion(job, nifi_xml)

        return job

    def _simulate_conversion(self, job: ConversionJob, nifi_xml: str):
        """
        Simulate the conversion process (placeholder for actual agent).

        In production, this should be replaced with actual agent invocation.
        """
        try:
            # Parse the XML to identify flows
            flows = NiFiParser.parse_nifi_xml(nifi_xml)

            job.status = JobStatus.PROCESSING
            job.started_at = datetime.now()
            job.status_message = f"Processing {len(flows)} flow(s)..."
            job.progress_percentage = 25

            # Simulate notebook generation for each flow
            notebooks = []
            for idx, flow in enumerate(flows):
                notebook_name = f"{flow.flow_name.replace(' ', '_').lower()}.py"

                # Generate placeholder notebook content
                notebook_content = self._generate_placeholder_notebook(flow)

                notebook = Notebook(
                    notebook_id=f"nb_{job.job_id}_{idx}",
                    notebook_name=notebook_name,
                    flow_id=flow.flow_id,
                    content=notebook_content,
                    language="python"
                )
                notebooks.append(notebook)

                # Update progress
                job.progress_percentage = 25 + (idx + 1) * (75 // len(flows))
                job.status_message = f"Generated notebook {idx + 1} of {len(flows)}"

            # Store notebook paths
            job.generated_notebooks = [nb.notebook_name for nb in notebooks]
            job.status = JobStatus.COMPLETED
            job.completed_at = datetime.now()
            job.progress_percentage = 100
            job.status_message = f"Successfully generated {len(notebooks)} notebook(s)"

        except Exception as e:
            job.status = JobStatus.FAILED
            job.completed_at = datetime.now()
            job.error_message = str(e)
            job.status_message = "Conversion failed"

    def _generate_placeholder_notebook(self, flow) -> str:
        """Generate placeholder notebook content."""
        return f'''# Databricks notebook source
# MAGIC %md
# MAGIC # {flow.flow_name}
# MAGIC
# MAGIC **Flow ID:** {flow.flow_id}
# MAGIC
# MAGIC **Description:** {flow.description or "No description available"}
# MAGIC
# MAGIC ---
# MAGIC
# MAGIC This notebook was generated from a NiFi flow using the agentic conversion system.

# COMMAND ----------

# MAGIC %md
# MAGIC ## Setup

# COMMAND ----------

from pyspark.sql import SparkSession
from pyspark.sql import functions as F
import logging

# Initialize logger
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

logger.info(f"Starting flow: {flow.flow_name}")

# COMMAND ----------

# MAGIC %md
# MAGIC ## Data Processing
# MAGIC
# MAGIC TODO: Implement the actual data processing logic based on NiFi processors

# COMMAND ----------

# Placeholder for data processing logic
# This will be replaced by actual implementation from the agent

logger.info("Processing data...")

# Example: Read data from source
# df = spark.read.format("delta").load("/path/to/source")

# Example: Transform data
# df_transformed = df.filter(F.col("status") == "active")

# Example: Write data to destination
# df_transformed.write.format("delta").mode("overwrite").save("/path/to/destination")

logger.info("Processing complete")

# COMMAND ----------

# MAGIC %md
# MAGIC ## Summary

# COMMAND ----------

logger.info(f"Flow {flow.flow_name} completed successfully")
'''

    def get_conversion_status(self, job_id: str) -> Optional[ConversionJob]:
        """Get the status of a conversion job."""
        return self.jobs.get(job_id)

    def get_conversion_results(self, job_id: str):
        """
        Get the conversion results (notebooks) for a job.

        Returns:
            List of Notebook objects
        """
        job = self.jobs.get(job_id)
        if not job:
            raise Exception(f"Job not found: {job_id}")

        if job.status != JobStatus.COMPLETED:
            raise Exception(f"Job not completed. Current status: {job.status.value}")

        # TODO: Replace with actual notebook retrieval from agent
        # For now, return empty list as notebooks are generated in-memory
        return []

    def cancel_conversion(self, job_id: str) -> bool:
        """Cancel a conversion job."""
        job = self.jobs.get(job_id)
        if not job:
            return False

        if job.status in [JobStatus.COMPLETED, JobStatus.FAILED, JobStatus.CANCELLED]:
            return False

        job.status = JobStatus.CANCELLED
        job.completed_at = datetime.now()
        job.status_message = "Job cancelled by user"
        return True
